requirements.txt
streamlit>=1.31.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
pandas>=2.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
plotly>=5.17.0
